refactor: Move chain configurations to separate chains.ts file
- Extracted SUPPORTED_CHAINS array and related types/functions to chains.ts
- Kept all chain configuration logic in a dedicated module
- Improved code organization and separation of concerns
- No functional changes, purely organizational refactoring
